remove ULONG/PULONG

+++ chakracore-cxx/lib/Backend/IRBuilder.cpp
@@ -37,7 +37,7 @@ IRBuilder::AddStatementBoundary(uint statementIndex, uint offset)
         {
             if (Js::Configuration::Global.flags.IsEnabled(Js::BailOutFlag) && !this->m_func->GetJITFunctionBody()->IsLibraryCode())
             {
-                ULONG line;
+                uint32_t line;
                 int32_t col;
 
                 // Since we're on a separate thread, don't allow the line cache to be allocated in the Recycler.
